Use template for rotl, fix shift by 0 places
With this template, data types other than 64-bit ulongs can be
bit-rotated, provided they are 1) unsigned, 2) arithmetic. This
eliminates the assumption that all ulongs are 64-bits in size, and
allows the function to potentially be used more generally.

More importantly, this also fixes (potentially?) undefined behaviour
when a rotate is performed by 0 places.

With GDC the code still compiles down to one instruction on x86-64.